requirements.txt
h5py==2.10.0
Keras-Preprocessing==1.1.2
matplotlib==3.2.0
numpy==1.19.3
opencv-python==4.5.1.48
Pillow==8.2.0
tensorboard==2.5.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.4.0
tensorflow-estimator==2.4.0
